+++ packages/treecharts/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./svgHelpers";
 export * from "./ConnectionDrawer";
 export * from "./NodeDrawer";
 export * from "./TitleDrawer";
+export * from "./ActionDrawer";